@@ -23,6 +23,7 @@ import (
 
 	"github.com/elastic/elastic-agent-libs/kibana"
 
+	atesting "github.com/elastic/elastic-agent/pkg/testing"
 	"github.com/elastic/elastic-agent/pkg/testing/define"
 	"github.com/elastic/elastic-agent/pkg/testing/tools"
 
@@ -92,7 +93,11 @@ func TestFQDN(t *testing.T) {
 			},
 		},
 	}
-	policy, err := tools.InstallAgentWithPolicy(t, agentFixture, kibClient, createPolicyReq)
+	installOpts := atesting.InstallOpts{
+		NonInteractive: true,
+		Force:          true,
+	}
+	policy, err := tools.InstallAgentWithPolicy(t, installOpts, agentFixture, kibClient, createPolicyReq)
 	require.NoError(t, err)
 
 	t.Log("Verify that agent name is short hostname")
@@ -162,9 +167,9 @@ func TestFQDN(t *testing.T) {
 
 	// TODO: Re-enable assertion once https://github.com/elastic/elastic-agent/issues/3078 is
 	// investigated for root cause and resolved.
-	//t.Log("Verify that hostname in `logs-*` and `metrics-*` is short hostname again")
-	//verifyHostNameInIndices(t, "logs-*", shortName, info.ESClient)
-	//verifyHostNameInIndices(t, "metrics-*", shortName, info.ESClient)
+	// t.Log("Verify that hostname in `logs-*` and `metrics-*` is short hostname again")
+	// verifyHostNameInIndices(t, "logs-*", shortName, info.ESClient)
+	// verifyHostNameInIndices(t, "metrics-*", shortName, info.ESClient)
 }
 
 func verifyAgentName(t *testing.T, hostname string, kibClient *kibana.Client) *kibana.AgentExisting {
@@ -276,7 +281,7 @@ func setHostFQDN(ctx context.Context, etcHosts []byte, externalIP, fqdn string, 
 	line := fmt.Sprintf("%s\t%s %s\n", externalIP, fqdn, shortName)
 
 	etcHosts = append(etcHosts, []byte(line)...)
-	err := os.WriteFile(filename, etcHosts, 0644)
+	err := os.WriteFile(filename, etcHosts, 0o644)
 	if err != nil {
 		return err
 	}
@@ -293,7 +298,7 @@ func setHostFQDN(ctx context.Context, etcHosts []byte, externalIP, fqdn string, 
 
 func setEtcHosts(data []byte) error {
 	filename := string(filepath.Separator) + filepath.Join("etc", "hosts")
-	return os.WriteFile(filename, data, 0644)
+	return os.WriteFile(filename, data, 0o644)
 }
 
 func setHostname(ctx context.Context, hostname string, log func(args ...any)) error {
